Removed out-pt dependency from join-manager
