+++ packages/geometry-utils/src/clipper/out-rec-manager.ts
@@ -4,6 +4,7 @@ import OutPt from "./out-pt";
 import OutRec from "./out-rec";
 import TEdge from "./t-edge";
 import { DIRECTION, NullPtr } from "./types";
+import { PointI32 } from "../geometry";
 
 export default class OutRecManager {
     private polyOuts: OutRec[] = [];
@@ -258,5 +259,153 @@ export default class OutRecManager {
             default:
                 return OutRec.getLowermostRec(outRec1, outRec2);
         }
+    }
+
+    public joinPoints(outHash1: number, outHash2: number, offPoint: Point<Int32Array>, isUseFullRange: boolean): { outHash1: number, outHash2: number, result: boolean  } {
+        const index1: number = get_u16_from_u32(outHash1, 0);
+        const index2: number = get_u16_from_u32(outHash2, 0);
+        const outRec1 = this.getOutRec(index1);
+        const outRec2 = this.getOutRec(index2);
+        const result = { outHash1, outHash2, result: false };
+        
+        if (outRec1.isEmpty || outRec2.isEmpty) {
+            return result;
+        }
+
+        const outPt1Index: number = get_u16_from_u32(outHash1, 1);
+        const outPt2Index: number = get_u16_from_u32(outHash2, 1);
+        let outPt1: OutPt = OutPt.getByIndex(outPt1Index);
+        let outPt2: OutPt = OutPt.getByIndex(outPt2Index);
+        const isRecordsSame = outRec1.index === outRec2.index;
+        //There are 3 kinds of joins for output polygons ...
+        //1. Horizontal joins where Join.OutPt1 & Join.OutPt2 are a vertices anywhere
+        //along (horizontal) collinear edges (& Join.OffPt is on the same horizontal).
+        //2. Non-horizontal joins where Join.OutPt1 & Join.OutPt2 are at the same
+        //location at the Bottom of the overlapping segment (& Join.OffPt is above).
+        //3. StrictlySimple joins where edges touch but are not collinear and where
+        //Join.OutPt1, Join.OutPt2 & Join.OffPt all share the same point.
+        const isHorizontal: boolean = outPt1.point.y === offPoint.y;
+
+        if (isHorizontal && offPoint.almostEqual(outPt1.point) && offPoint.almostEqual(outPt2.point)) {
+            //Strictly Simple join ...
+            const op1b = outPt1.strictlySimpleJoin(offPoint);
+            const op2b = outPt2.strictlySimpleJoin(offPoint);
+
+            const reverse1: boolean = op1b.point.y > offPoint.y;
+            const reverse2: boolean = op2b.point.y > offPoint.y;
+
+            if (reverse1 === reverse2) {
+                return result;
+            }
+
+            result.outHash2 = join_u16_to_u32(index2, outPt1.applyJoin(outPt2, reverse1).index);
+            result.result = true;
+
+            return result;
+        } else if (isHorizontal) {
+            //treat horizontal joins differently to non-horizontal joins since with
+            //them we're not yet sure where the overlapping is. OutPt1.Pt & OutPt2.Pt
+            //may be anywhere along the horizontal edge.
+            const outPt1Res = outPt1.flatHorizontal(outPt2, outPt2);
+
+            if (outPt1Res.length === 0) {
+                return result;
+            }
+
+            const [op1, op1b] = outPt1Res;
+            //a flat 'polygon'
+            const outPt2Res = outPt2.flatHorizontal(op1, op1b);
+
+            if (outPt2Res.length === 0) {
+                return result;
+            }
+
+            const [op2, op2b] = outPt2Res;
+            //a flat 'polygon'
+            //Op1 -. Op1b & Op2 -. Op2b are the extremites of the horizontal edges
+
+            const value = PointI32.getOverlap(op1.point.x, op1b.point.x, op2.point.x, op2b.point.x);
+            const isOverlapped = value.x < value.y;
+
+            if (!isOverlapped) {
+                return result;
+            }
+
+            //DiscardLeftSide: when overlapping edges are joined, a spike will created
+            //which needs to be cleaned up. However, we don't want Op1 or Op2 caught up
+            //on the discard Side as either may still be needed for other joins ...
+            const point = PointI32.create();
+            let discardLeftSide: boolean = false;
+            if (op1.point.x >= value.x && op1.point.x <= value.y) {
+                //Pt = op1.Pt;
+                point.update(op1.point);
+                discardLeftSide = op1.point.x > op1b.point.x;
+            } else if (op2.point.x >= value.x && op2.point.x <= value.y) {
+                //Pt = op2.Pt;
+                point.update(op2.point);
+                discardLeftSide = op2.point.x > op2b.point.x;
+            } else if (op1b.point.x >= value.x && op1b.point.x <= value.y) {
+                //Pt = op1b.Pt;
+                point.update(op1b.point);
+                discardLeftSide = op1b.point.x > op1.point.x;
+            } else {
+                //Pt = op2b.Pt;
+                point.update(op2b.point);
+                discardLeftSide = op2b.point.x > op2.point.x;
+            }
+            result.outHash1 = join_u16_to_u32(index1, op1.index);
+            result.outHash2 = join_u16_to_u32(index2, op2.index);
+            result.result = OutPt.joinHorz(op1, op1b, op2, op2b, point, discardLeftSide);
+
+            return result;
+        } else {
+            let op1 = outPt1;
+            let op2 = outPt2;
+            let op1b: OutPt = op1.getUniquePt(true);
+            let op2b: OutPt = op2.getUniquePt(true);
+            //nb: For non-horizontal joins ...
+            //    1. Jr.OutPt1.Pt.Y === Jr.OutPt2.Pt.Y
+            //    2. Jr.OutPt1.Pt > Jr.OffPt.Y
+            //make sure the polygons are correctly oriented ...
+
+            const reverse1: boolean =
+                op1b.point.y > op1.point.y || !PointI32.slopesEqual(op1.point, op1b.point, offPoint, isUseFullRange);
+
+            if (reverse1) {
+                op1b = op1.getUniquePt(false);
+
+                if (op1b.point.y > op1.point.y || !PointI32.slopesEqual(op1.point, op1b.point, offPoint, isUseFullRange)) {
+                    return result;
+                }
+            }
+
+            const reverse2: boolean =
+                op2b.point.y > op2.point.y || !PointI32.slopesEqual(op2.point, op2b.point, offPoint, isUseFullRange);
+
+            if (reverse2) {
+                op2b = op2.getUniquePt(false);
+
+                if (op2b.point.y > op2.point.y || !PointI32.slopesEqual(op2.point, op2b.point, offPoint, isUseFullRange)) {
+                    return result;
+                }
+            }
+
+            if (op1b === op1 || op2b === op2 || op1b === op2b || (isRecordsSame && reverse1 === reverse2)) {
+                return result;
+            }
+
+            result.outHash2 = join_u16_to_u32(index2, outPt1.applyJoin(outPt2, reverse1).index);
+
+            result.result = true;
+
+            return result;
+        }
+    }
+
+    public horzSegmentsOverlap(outHash: number, offPoint: Point<Int32Array>, rightBound: TEdge): boolean {
+        const outPtIndex = get_u16_from_u32(outHash, 1);
+        const outPt = OutPt.getByIndex(outPtIndex);
+        
+        return PointI32.horzSegmentsOverlap(outPt.point, offPoint, rightBound.Bot, rightBound.Top);
     }
 }
